Initial commit [...]
Initial commit of the new TurboTools (just TurboShell for now), which is a reboot/rewrite of an unnamed-software that I've been co-developing under various contracts for the past few years.

This new TurboTools will be written in AutoIt - an excellent and extremely powerful scripting language for RAD and automation - and will remain 100% opensource and patent-free - albeit with the conditions of all associated AutoIt interpreter/contributor terms.